pyproject.toml
[project]
name = "geo-dataset-builder"
version = "0.1.0"
description = ""
authors = [
    {name = "gaetan.tiazara",email = "gaetan.tiazara@ign.fr"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas (>=1.1.1,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "mercantile (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "h5py (>=3.15.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "geo_dataset_builder", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
